requirements.txt
pandas==2.2.2
Django==5.1
djangorestframework==3.15.2
python-dotenv==1.0.1